requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.6
exceptiongroup==1.1.2
Flask==2.3.2
Flask-RESTful==0.3.10
flask-swagger-ui==4.11.1
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
tomli==2.0.1
Werkzeug==2.3.6
zipp==3.16.2
